+++ Example/DApp/AppDelegate.swift
@@ -4,6 +4,16 @@ import UIKit
 class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         // Override point for customization after application launch.
+        
+        #if DEBUG
+        if ProcessInfo.processInfo.arguments.contains("-disableAnimations") {
+            UIView.setAnimationsEnabled(false)
+            UIApplication.shared.connectedScenes
+                .flatMap { ($0 as? UIWindowScene)?.windows ?? [] }
+                .last { $0.isKeyWindow }?.layer.speed = 200
+        }
+        #endif
+        
         return true
     }
